@@ -83,82 +82,78 @@
     }
   },
   "engines": {
-    "node": ">=10 <=15"
+    "node": ">=12.20.2 <=19"
   },
   "dependencies": {
     "commander": "2.20.0",
     "electron-window-state": "5.0.3",
     "post-robot": "https://github.com/dragonbane0/post-robot/tarball/dad59c1c0f6266d0dd4bd35e0fb7137d45c6c773",
-    "tree-kill": "1.2.1"
+    "tree-kill": "1.2.1",
+    "tslib": "2.0.0"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "~0.803.29",
-    "@angular/animations": "8.2.14",
-    "@angular/cdk": "8.2.3",
-    "@angular/cli": "8.3.29",
-    "@angular/common": "8.2.14",
-    "@angular/compiler": "8.2.14",
-    "@angular/compiler-cli": "8.2.14",
-    "@angular/core": "8.2.14",
-    "@angular/flex-layout": "8.0.0-beta.27",
-    "@angular/forms": "8.2.14",
-    "@angular/language-service": "8.2.14",
-    "@angular/material": "8.2.3",
-    "@angular/platform-browser": "8.2.14",
-    "@angular/platform-browser-dynamic": "8.2.14",
-    "@angular/router": "8.2.14",
-    "@angularclass/hmr": "2.1.3",
-    "@fortawesome/fontawesome-free": "5.9.0",
-    "@nebular/eva-icons": "4.6.0",
-    "@nebular/theme": "4.6.0",
-    "@types/electron": "1.6.10",
-    "@types/jasmine": "2.8.16",
+    "@angular-devkit/build-angular": "13.3.9",
+    "@angular-eslint/builder": "13.5.0",
+    "@angular-eslint/eslint-plugin": "13.5.0",
+    "@angular-eslint/eslint-plugin-template": "13.5.0",
+    "@angular-eslint/schematics": "13.5.0",
+    "@angular-eslint/template-parser": "13.5.0",
+    "@angular/animations": "13.3.11",
+    "@angular/cdk": "13.3.9",
+    "@angular/cli": "13.3.9",
+    "@angular/common": "13.3.11",
+    "@angular/compiler": "13.3.11",
+    "@angular/compiler-cli": "13.3.11",
+    "@angular/core": "13.3.11",
+    "@angular/flex-layout": "13.0.0-beta.38",
+    "@angular/forms": "13.3.11",
+    "@angular/language-service": "13.3.11",
+    "@angular/material": "13.3.9",
+    "@angular/platform-browser": "13.3.11",
+    "@angular/platform-browser-dynamic": "13.3.11",
+    "@angular/router": "13.3.11",
+    "@angularclass/hmr": "3.0.0",
+    "@fortawesome/fontawesome-free": "6.2.0",
+    "@nebular/eva-icons": "9.1.0-rc.6",
+    "@nebular/theme": "9.1.0-rc.6",
+    "@types/jasmine": "3.6.0",
     "@types/jasminewd2": "2.0.6",
-    "@types/node": "8.9.5",
+    "@types/node": "12.11.1",
+    "@typescript-eslint/eslint-plugin": "5.27.1",
+    "@typescript-eslint/parser": "5.27.1",
     "angular-dual-listbox": "4.7.0",
-    "angular-file": "1.3.2",
-    "browserslist": "^4.21.4",
-    "codelyzer": "^5.0.1",
+    "angular-file": "3.6.0",
+    "browserslist": "4.21.4",
     "core-js": "3.25.1",
     "crc": "3.8.0",
-    "electron": "4.2.6",
+    "electron": "7.3.3",
+    "eslint": "8.22.0",
     "eva-icons": "1.1.3",
     "hammerjs": "2.0.8",
     "intl": "1.2.5",
     "ionicons": "4.6.3",
-    "jasmine-core": "2.99.1",
-    "jasmine-spec-reporter": "4.2.1",
+    "jasmine-core": "3.8.0",
+    "jasmine-spec-reporter": "5.0.0",
     "jquery": "3.4.1",
-    "karma": "3.1.4",
-    "karma-chrome-launcher": "2.2.0",
+    "karma": "6.4.1",
+    "karma-chrome-launcher": "3.1.0",
     "karma-coverage-istanbul-reporter": "2.0.5",
-    "karma-jasmine": "1.1.2",
-    "karma-jasmine-html-reporter": "0.2.2",
+    "karma-jasmine": "4.0.0",
+    "karma-jasmine-html-reporter": "1.5.0",
     "ngx-color-picker": "https://github.com/dragonbane0/ngx-color-picker/tarball/7471e8595e5bf508813c5f6b1964e53ac9902448",
-    "node-sass": "4.14.1",
     "normalize.css": "6.0.0",
-    "pace-progress": "https://github.com/dragonbane0/pace/tarball/bcbd25563bf6193b223f856dc1668c0705fd8710",
-    "popper.js": "1.16.1",
-    "protractor": "5.4.2",
+    "pace-js": "1.2.4",
+    "protractor": "7.0.0",
     "roboto-fontface": "0.8.0",
     "rxjs": "6.6.7",
     "rxjs-compat": "6.3.0",
+    "sass": "1.55.0",
     "socicon": "3.0.5",
     "tcp-ping": "0.1.1",
     "ts-node": "7.0.1",
-    "tslib": "1.10.0",
-    "tslint": "5.11.0",
     "typeface-exo": "0.0.22",
-    "typescript": "3.5.3",
-    "web-animations-js": "^2.3.2",
-    "zone.js": "~0.9.1"
+    "typescript": "4.6.4",
+    "zone.js": "0.11.4"
   },
-  "overrides": {
-    "browserslist": {
-      "caniuse-lite": "1.0.30001393"
-    },
-    "autoprefixer": {
-      "caniuse-lite": "1.0.30001393"
-    }
-  }
+  "overrides": {}
 }
