@@ -63,10 +63,10 @@ function romBuilding(pythonPath, randoPath, settings) {
       let args = ['--seed', seedString];
       //args["checked_version"] = false;
 
-      romBuildingGenerator = spawn('"' + pythonPath + '"' + ' ' + '"' + randoPath + '"', args, { shell: true });
+      romBuildingGenerator = spawn(pythonPath + ' ' + '"' + randoPath + '"', args, { shell: true });
     }
     else { //Random Seed
-      romBuildingGenerator = spawn('"' + pythonPath + '"' + ' ' + '"' + randoPath + '"', { shell: true });
+      romBuildingGenerator = spawn(pythonPath + ' ' + '"' + randoPath + '"', { shell: true });
     }
 
     romBuildingGenerator.on('error', err => {
@@ -232,7 +232,7 @@ function testPythonPath(pythonPath) {
 
     var error = "";
 
-    let pythonExec = spawn('"' + pythonPath + '"', { shell: true }).on('error', err => {
+    let pythonExec = spawn(pythonPath, { shell: true }).on('error', err => {
       reject(err);
     });
 
@@ -270,7 +270,7 @@ function getSettings(pythonPath, randoPath, settingsString) {
 
     //console.log("Get settings now with spawn!");
 
-    let settingsPY = spawn('"' + pythonPath + '"' + ' ' + '"' + randoPath + '"', args, { shell: true }).on('error', err => {
+    let settingsPY = spawn(pythonPath + ' ' + '"' + randoPath + '"', args, { shell: true }).on('error', err => {
       console.error("[getSettings] Error spawning process:", err);
       reject(err);
     });
@@ -314,7 +314,7 @@ function parseSettings(pythonPath, randoPath) {
 
     let args = ['--convert_settings'];
 
-    let settingsPY = spawn('"' + pythonPath + '"' + ' ' + '"' + randoPath + '"', args, { shell: true }).on('error', err => {
+    let settingsPY = spawn(pythonPath + ' ' + '"' + randoPath + '"', args, { shell: true }).on('error', err => {
       console.error("[parseSettings] Error spawning process:", err);
       reject(err);
     });
