@@ -1,11 +1,11 @@
-import { Component } from '@angular/core';
+import { Component, OnInit } from '@angular/core';
 import { NbDialogService } from '@nebular/theme';
 import { GUIGlobal } from '../../../providers/GUIGlobal';
 
-import { ConfirmationWindow } from '../../../pages/generator/confirmationWindow/confirmationWindow.component';
+import { ConfirmationWindowComponent } from '../../../pages/generator/confirmationWindow/confirmationWindow.component';
 
 @Component({
-  selector: 'ngx-footer',
+  selector: 'ootr-footer',
   styleUrls: ['./footer.component.scss'],
   template: `
     <span *ngIf="!hasUpdate">Version <b>{{localVersion}}</b> © ZeldaSpeedRuns Community</span>
@@ -20,7 +20,7 @@ import { ConfirmationWindow } from '../../../pages/generator/confirmationWindow/
     </div>
   `,
 })
-export class FooterComponent {
+export class FooterComponent implements OnInit {
 
   localVersion: string = "";
   remoteVersion: string = "";
@@ -62,7 +62,7 @@ export class FooterComponent {
   }
 
   promptUpdate() {
-    this.dialogService.open(ConfirmationWindow, {
+    this.dialogService.open(ConfirmationWindowComponent, {
       autoFocus: true, closeOnBackdropClick: true, closeOnEsc: true, hasBackdrop: true, hasScroll: false, context: { dialogHeader: "New Version Available!", dialogMessage: "You are using version " + this.localVersion + ", and the latest is version " + this.remoteVersion + ". Do you want to download the latest version now?" + ((this.remoteVersion.includes("Release")) ? "" : " (Note that you are using a development build and therefore will have to redownload and compile the source off GitHub yourself)") }
     }).onClose.subscribe(confirmed => {
 
