@@ -8,8 +8,9 @@ const routes: Routes = [
 ];
 
 const config: ExtraOptions = {
-  useHash: true,
-  initialNavigation: false //Route manually at the start to avoid URL changes in the browser
+    useHash: true,
+    initialNavigation: 'disabled', //Route manually at the start to avoid URL changes in the browser
+    relativeLinkResolution: 'legacy'
 };
 
 @NgModule({
