@@ -1,9 +1,9 @@
 import { APP_BASE_HREF } from '@angular/common';
 import { BrowserModule } from '@angular/platform-browser';
 import { BrowserAnimationsModule } from '@angular/platform-browser/animations';
-import { NgModule, ComponentFactoryResolver, ApplicationRef, CUSTOM_ELEMENTS_SCHEMA } from '@angular/core';
+import { NgModule, ComponentFactoryResolver, ApplicationRef, CUSTOM_ELEMENTS_SCHEMA, DoBootstrap } from '@angular/core';
 import { NbDialogModule } from '@nebular/theme';
-import { NbEvaIconsModule } from "@nebular/eva-icons";
+import { NbEvaIconsModule } from '@nebular/eva-icons';
 
 import { HttpClientModule } from '@angular/common/http';
 
@@ -35,70 +35,62 @@ import { MatTableModule } from '@angular/material/table';
 import { GUIGlobal } from './providers/GUIGlobal';
 
 //Custom Components
-import { ProgressWindow } from './pages/generator/progressWindow/progressWindow.component';
-import { DialogWindow } from './pages/generator/dialogWindow/dialogWindow.component';
-import { ErrorDetailsWindow } from './pages/generator/errorDetailsWindow/errorDetailsWindow.component';
-import { ConfirmationWindow } from './pages/generator/confirmationWindow/confirmationWindow.component';
-import { TextInputWindow } from './pages/generator/textInputWindow/textInputWindow.component';
+import { ProgressWindowComponent } from './pages/generator/progressWindow/progressWindow.component';
+import { DialogWindowComponent } from './pages/generator/dialogWindow/dialogWindow.component';
+import { ErrorDetailsWindowComponent } from './pages/generator/errorDetailsWindow/errorDetailsWindow.component';
+import { ConfirmationWindowComponent } from './pages/generator/confirmationWindow/confirmationWindow.component';
+import { TextInputWindowComponent } from './pages/generator/textInputWindow/textInputWindow.component';
 
 
 @NgModule({
-  declarations: [
-    AppComponent,
-    BypassSecurityPipe,
-    ProgressWindow,
-    DialogWindow,
-    ErrorDetailsWindow,
-    ConfirmationWindow,
-    TextInputWindow
-  ],
-  entryComponents: [
-    AppComponent,
-    ProgressWindow,
-    DialogWindow,
-    ErrorDetailsWindow,
-    ConfirmationWindow,
-    TextInputWindow
-  ],
-  imports: [
-    BrowserModule,
-    BrowserAnimationsModule,
-    AppRoutingModule,
-    HttpClientModule,
-    FormsModule,
-    ReactiveFormsModule,
-    CdkTableModule,
-    ThemeModule.forRoot(),
-    NbDialogModule.forRoot(),
-    NbEvaIconsModule
-  ],
-  exports: [
-    MatButtonModule,
-    MatButtonToggleModule,
-    MatCardModule,
-    MatCheckboxModule,
-    MatDialogModule,
-    MatGridListModule,
-    MatIconModule,
-    MatInputModule,
-    MatListModule,
-    MatProgressBarModule,
-    MatProgressSpinnerModule,
-    MatRadioModule,
-    MatSelectModule,
-    MatSliderModule,
-    MatSlideToggleModule,
-    MatTableModule
-  ],
-  schemas: [
-    CUSTOM_ELEMENTS_SCHEMA
-  ],
-  providers: [
-    { provide: APP_BASE_HREF, useValue: '/' },
-    GUIGlobal
-  ]
+    declarations: [
+        AppComponent,
+        BypassSecurityPipe,
+        ProgressWindowComponent,
+        DialogWindowComponent,
+        ErrorDetailsWindowComponent,
+        ConfirmationWindowComponent,
+        TextInputWindowComponent
+    ],
+    imports: [
+        BrowserModule,
+        BrowserAnimationsModule,
+        AppRoutingModule,
+        HttpClientModule,
+        FormsModule,
+        ReactiveFormsModule,
+        CdkTableModule,
+        ThemeModule.forRoot(),
+        NbDialogModule.forRoot(),
+        NbEvaIconsModule
+    ],
+    exports: [
+        MatButtonModule,
+        MatButtonToggleModule,
+        MatCardModule,
+        MatCheckboxModule,
+        MatDialogModule,
+        MatGridListModule,
+        MatIconModule,
+        MatInputModule,
+        MatListModule,
+        MatProgressBarModule,
+        MatProgressSpinnerModule,
+        MatRadioModule,
+        MatSelectModule,
+        MatSliderModule,
+        MatSlideToggleModule,
+        MatTableModule
+    ],
+    schemas: [
+        CUSTOM_ELEMENTS_SCHEMA
+    ],
+    providers: [
+        { provide: APP_BASE_HREF, useValue: '/' },
+        GUIGlobal
+    ]
 })
-export class AppModule {
+export class AppModule implements DoBootstrap {
 
   constructor(private resolver: ComponentFactoryResolver) { }
 
