@@ -23,51 +23,47 @@ import { MatSliderModule } from '@angular/material/slider';
 import { MatTableModule } from '@angular/material/table';
 
 import { AngularDualListBoxModule } from 'angular-dual-listbox';
-import { GUIListboxModule } from '../../components/guiListbox/guiListbox';
+import { GUIListboxComponent } from '../../components/guiListbox/guiListbox';
 import { ColorPickerModule } from 'ngx-color-picker';
-import { ngfModule } from "angular-file";
+import { ngfModule } from 'angular-file';
 
 //Custom Directives
 import { ResponsiveColsDirective } from '../../directives/responsiveCols.directive';
 
 //Custom Components
-import { GUITooltip } from './guiTooltip/guiTooltip.component';
+import { GUITooltipComponent } from './guiTooltip/guiTooltip.component';
 
 @NgModule({
-  imports: [
-    ThemeModule,
-    MatButtonModule,
-    MatButtonToggleModule,
-    MatCardModule,
-    MatCheckboxModule,
-    MatDialogModule,
-    MatGridListModule,
-    MatIconModule,
-    MatInputModule,
-    MatListModule,
-    MatProgressBarModule,
-    MatProgressSpinnerModule,
-    MatRadioModule,
-    MatSelectModule,
-    MatSliderModule,
-    MatSlideToggleModule,
-    MatTableModule,
-    FlexLayoutModule,
-    ColorPickerModule,
-    ngfModule
-  ],
-  declarations: [
-    GeneratorComponent,
-    ResponsiveColsDirective,
-    GUITooltip,
-    GUIListboxModule
-  ],
-  entryComponents: [
-    GUITooltip,
-    GUIListboxModule
-  ],
-  schemas: [
-    CUSTOM_ELEMENTS_SCHEMA
-  ]
+    imports: [
+        ThemeModule,
+        MatButtonModule,
+        MatButtonToggleModule,
+        MatCardModule,
+        MatCheckboxModule,
+        MatDialogModule,
+        MatGridListModule,
+        MatIconModule,
+        MatInputModule,
+        MatListModule,
+        MatProgressBarModule,
+        MatProgressSpinnerModule,
+        MatRadioModule,
+        MatSelectModule,
+        MatSliderModule,
+        MatSlideToggleModule,
+        MatTableModule,
+        FlexLayoutModule,
+        ColorPickerModule,
+        ngfModule
+    ],
+    declarations: [
+        GeneratorComponent,
+        ResponsiveColsDirective,
+        GUITooltipComponent,
+        GUIListboxComponent
+    ],
+    schemas: [
+        CUSTOM_ELEMENTS_SCHEMA
+    ]
 })
 export class GeneratorModule { }
